+++ source/Public/Update-PstProject.ps1
@@ -25,7 +25,11 @@ function Update-PstProject {
         [Parameter(Mandatory)]
         [ArgumentCompleter({
             param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameters)
-            $configFile = Join-Path $env:localappdata 'PstTimeTracker\config.json'
+            if ($IsWindows -or $PSVersionTable.PSVersion.Major -le 5) {
+                $configFile = Join-Path $env:LOCALAPPDATA 'PstTimeTracker\config.json'
+            } else {
+                $configFile = Join-Path $HOME '.local/share/PstTimeTracker/config.json'
+            }
             if (Test-Path $configFile) {
                 $config = Get-Content $configFile -Raw | ConvertFrom-Json
                 $config.Clients.Name | Where-Object { $_ -like "$wordToComplete*" }
@@ -38,7 +42,11 @@ function Update-PstProject {
             param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameters)
             $clientName = $fakeBoundParameters['Client']
             if ($clientName) {
-                $configFile = Join-Path $env:localappdata 'PstTimeTracker\config.json'
+                if ($IsWindows -or $PSVersionTable.PSVersion.Major -le 5) {
+                    $configFile = Join-Path $env:LOCALAPPDATA 'PstTimeTracker\config.json'
+                } else {
+                    $configFile = Join-Path $HOME '.local/share/PstTimeTracker/config.json'
+                }
                 if (Test-Path $configFile) {
                     $config = Get-Content $configFile -Raw | ConvertFrom-Json
                     $client = $config.Clients | Where-Object { $_.Name -eq $clientName }